+++ CHANGES.rst
@@ -9,6 +9,8 @@ API Changes
 
 Bug Fixes
 ^^^^^^^^^
+- When all-zero bin encountered in fit_trace with peak_method=gaussian, the bin peak
+  will be set to NaN in this caseto work better with DogBoxLSQFitter. [#257]
 
 Other changes
 ^^^^^^^^^^^^^